Swap around the versions now that we are going to release a 9.0 major version.
Update to latest master branch.
