+++ test/HealthChecks.RabbitMQ.Tests/HealthChecks.RabbitMQ.Tests.csproj
@@ -1,8 +1,4 @@
 <Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <DefineConstants>$(DefineConstants);RABBITMQ_V6</DefineConstants>
-  </PropertyGroup>
   
   <ItemGroup>
     <ProjectReference Include="..\..\src\HealthChecks.Rabbitmq\HealthChecks.Rabbitmq.csproj" />
