+++ test/HealthChecks.RabbitMQ.v6.Tests/HealthChecks.Rabbitmq.approved.txt
